requirements.txt
# Django 1.6.x
Django<=1.7
pillow
-e git+https://github.com/point97/django-mbtiles.git#egg=django-mbtiles
django-tastypie
